create data schema file
