+++ app/main.py
@@ -1,24 +1,16 @@
-import re
 import mimetypes
-
-from app.config import Config, Gen3Config, iRODSConfig
-
 from fastapi import FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import PlainTextResponse, StreamingResponse, JSONResponse, Response
-
-from typing import Union
-from pydantic import BaseModel
-from enum import Enum
-
 from gen3.auth import Gen3Auth
 from gen3.submission import Gen3Submission
-
-from app.sgqlc import SimpleGraphQLClient
-from app.filter import Filter, FIELDS
-
 from irods.session import iRODSSession
-from app.search import Search
+
+from app.config import Config, Gen3Config, iRODSConfig
+from app.data_schema import *
+from app.sgqlc import SimpleGraphQLClient
+from app.filter import Filter
+from app.pagination import Pagination
 
 description = """
 ## Gen3
@@ -61,16 +53,16 @@ tags_metadata = [
 app = FastAPI(
     title="12 Labours Portal",
     description=description,
-    # version="0.0.1",
-    # terms_of_service="http://example.com/terms/",
+    # version="",
+    # terms_of_service="",
     contact={
         "name": "Auckland Bioengineering Institute",
         "url": "https://www.auckland.ac.nz/en/abi.html",
-        "email": "bioeng-enquiries@auckland.ac.nz",
+        # "email": "bioeng-enquiries@auckland.ac.nz",
     },
     # license_info={
-    #     "name": "Apache 2.0",
-    #     "url": "https://www.apache.org/licenses/LICENSE-2.0.html",
+    #     "name": "",
+    #     "url": "",
     # }
     openapi_tags=tags_metadata
 )
@@ -89,19 +81,12 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-BAD_REQUEST = 400
-UNAUTHORIZED = 401
-NOT_FOUND = 404
-METHOD_NOT_ALLOWED = 405
-INTERNAL_SERVER_ERROR = 500
-
-
 SUBMISSION = None
 SESSION = None
 
 sgqlc = SimpleGraphQLClient()
 f = Filter()
-s = Search()
+p = Pagination()
 
 
 @ app.on_event("startup")
@@ -140,6 +125,12 @@ async def root():
 # Gen3 Data Commons
 #
 
+def get_name_list(data, name, path):
+    name_dict = {name: []}
+    for ele in data["links"]:
+        name_dict[name].append(ele.replace(path, ""))
+    return name_dict
+
 
 @ app.get("/program", tags=["Gen3"])
 async def get_gen3_program():
@@ -148,20 +139,13 @@ async def get_gen3_program():
     """
     try:
         program = SUBMISSION.get_programs()
-        program_dict = {"program": []}
-        for ele in program["links"]:
-            program_dict["program"].append(ele.replace("/v0/submission/", ""))
-        return program_dict
+        return get_name_list(program, "program", "/v0/submission/")
     except Exception as e:
         raise HTTPException(status_code=NOT_FOUND, detail=str(e))
 
 
-class Program(str, Enum):
-    program = "demo1"
-
-
-@ app.get("/project/{program}", tags=["Gen3"])
-async def get_gen3_project(program: Program):
+@ app.get("/project/{program}", tags=["Gen3"], summary="Get gen3 project information", response_description="Gen3 project name")
+async def get_gen3_project(program: ProgramParam):
     """
     Return all projects information from a program.
 
@@ -169,29 +153,12 @@ async def get_gen3_project(program: Program):
     """
     try:
         project = SUBMISSION.get_projects(program)
-        project_dict = {"project": []}
-        for ele in project["links"]:
-            project_dict["project"].append(
-                ele.replace(f"/v0/submission/{program}/", ""))
-        return project_dict
+        return get_name_list(project, "project", f"/v0/submission/{program}/")
     except Exception as e:
         raise HTTPException(status_code=NOT_FOUND, detail=str(e))
 
 
-class Gen3Item(BaseModel):
-    program: Union[str, None] = None
-    project: Union[str, None] = None
-
-    class Config:
-        schema_extra = {
-            "example": {
-                "program": "demo1",
-                "project": "12L",
-            }
-        }
-
-
-@ app.post("/dictionary", tags=["Gen3"])
+@ app.post("/dictionary", tags=["Gen3"], summary="Get gen3 dictionary information", response_description="Gen3 dictionary name")
 async def get_gen3_dictionary(item: Gen3Item):
     """
     Return all dictionary nodes from the Gen3 Data Commons
@@ -203,24 +170,14 @@ async def get_gen3_dictionary(item: Gen3Item):
     try:
         dictionary = SUBMISSION.get_project_dictionary(
             item.program, item.project)
-        dictionary_dict = {"dictionary": []}
-        for ele in dictionary["links"]:
-            dictionary_dict["dictionary"].append(ele.replace(
-                f"/v0/submission/{item.program}/{item.project}/_dictionary/", ""))
-        return dictionary_dict
+        return get_name_list(dictionary, "dictionary", f"/v0/submission/{item.program}/{item.project}/_dictionary/")
     except Exception:
         raise HTTPException(
             status_code=NOT_FOUND, detail=f"Program {item.program} or project {item.project} not found")
 
 
-class Node(str, Enum):
-    experiment = "experiment"
-    dataset_description = "dataset_description"
-    manifest = "manifest"
-
-
-@ app.post("/records/{node}", tags=["Gen3"])
-async def get_gen3_node_records(node: Node, item: Gen3Item):
+@ app.post("/records/{node}", tags=["Gen3"], summary="Get gen3 node records information", response_description="A list of json object contains all records metadata within a node")
+async def get_gen3_node_records(node: NodeParam, item: Gen3Item):
     """
     Return all records information in a dictionary node.
 
@@ -270,28 +227,7 @@ async def get_gen3_record(uuid: str, item: Gen3Item):
         return record
 
 
-class GraphQLQueryItem(BaseModel):
-    page: Union[int, None] = 1
-    limit: Union[int, None] = 0
-    node: Union[str, None] = None
-    filter: Union[dict, None] = {}
-    search: Union[str, None] = ""
-
-    class Config:
-        schema_extra = {
-            "example": {
-                "node": "dataset_description",
-                "filter": {
-                    "submitter_id": [
-                        "dataset-<dataset_id>-version-<version_id>-dataset_description"
-                    ]
-                },
-                "search": "",
-            }
-        }
-
-
-@ app.post("/graphql/query", tags=["Gen3"])
+@ app.post("/graphql/query", tags=["Gen3"], summary="GraphQL query gen3 information")
 async def graphql_query(item: GraphQLQueryItem):
     """
     Return queries metadata records. The API uses GraphQL query language.
@@ -312,59 +248,7 @@ async def graphql_query(item: GraphQLQueryItem):
     return query_result[item.node]
 
 
-def update_pagination_item(item, input):
-    if item.filter != {}:
-        query_item = GraphQLQueryItem()
-        filter_dict = {"submitter_id": []}
-        temp_node_dict = {}
-        for element in item.filter.values():
-            query_item.node, query_item.filter = element["node"], element["filter"]
-            filter_node = re.sub('_filter', '', query_item.node)
-            filter_field = list(query_item.filter.keys())[0]
-            # Only do fetch when there is no related temp data stored in temp_node_dict
-            # or the node field type is "String"
-            if filter_node not in temp_node_dict.keys() or filter_field not in FIELDS:
-                query_result = sgqlc.get_queried_result(query_item, SUBMISSION)
-                # The data will be stored when the field type is an "Array"
-                # The default filter relation of the Gen3 "Array" type field is "AND"
-                # We need "OR", therefore entire node data will go through a self-written filter function
-                if filter_field in FIELDS:
-                    temp_node_dict[filter_node] = query_result[filter_node]
-            elif filter_node in temp_node_dict.keys() and filter_field in FIELDS:
-                query_result = temp_node_dict
-            filter_dict["submitter_id"].append(f.get_filtered_datasets(
-                query_item.filter, query_result[filter_node]))
-        item.filter = filter_dict
-        f.filter_relation(item)
-
-    if input != "":
-        # If input does not match any content in the database, item.search will be empty dictionary
-        item.search["submitter_id"] = s.get_searched_datasets(input, SESSION)
-        if item.search != {} and ("submitter_id" not in item.filter or item.filter["submitter_id"] != []):
-            s.search_filter_relation(item)
-
-
-class GraphQLPaginationItem(BaseModel):
-    page: Union[int, None] = 1
-    limit: Union[int, None] = 50
-    node: Union[str, None] = None
-    filter: Union[dict, None] = {}
-    search: Union[dict, None] = {}
-    relation: Union[str, None] = "and"
-
-    class Config:
-        schema_extra = {
-            "example": {
-                "page": 1,
-                "limit": 50,
-                "node": "experiment",
-                "filter": {},
-                "relation": "and"
-            }
-        }
-
-
-@ app.post("/graphql/pagination/", tags=["Gen3"])
+@ app.post("/graphql/pagination/", tags=["Gen3"], summary="Display datasets", response_description="A list of datasets")
 async def graphql_pagination(item: GraphQLPaginationItem, search: str = ""):
     """
     /graphql/pagination/?search=<string>
@@ -392,14 +276,14 @@ async def graphql_pagination(item: GraphQLPaginationItem, search: str = ""):
 
     :param search: string content.
     """
-    update_pagination_item(item, search)
+    p.update_pagination_item(item, search, SUBMISSION, SESSION)
     query_result = sgqlc.get_queried_result(item, SUBMISSION)
     if item.search != {}:
         # Sort only if search is not empty, since search results are sorted by word relevance
         query_result[item.node] = sorted(
             query_result[item.node], key=lambda dict: item.filter["submitter_id"].index(dict["submitter_id"]))
     return {
-        "data": query_result[item.node],
+        "items": p.update_pagination_output(query_result[item.node]),
         # Maximum number of records display in one page
         "limit": item.limit,
         "page": item.page,
@@ -415,17 +299,8 @@ async def generate_filter():
     return f.generate_filter_information()
 
 
-class Project(str, Enum):
-    project = "12L"
-
-
-class Format(str, Enum):
-    json = "json"
-    tsv = "tsv"
-
-
-@ app.get("/metadata/download/{program}/{project}/{uuid}/{format}", tags=["Gen3"])
-async def download_gen3_metadata_file(program: Program, project: Project, uuid: str, format: Format):
+@ app.get("/metadata/download/{program}/{project}/{uuid}/{format}", tags=["Gen3"], summary="Download gen3 record information", response_description="A JSON or CSV file contains the metadata")
+async def download_gen3_metadata_file(program: ProgramParam, project: ProjectParam, uuid: str, format: FormatParam):
     """
     Return a single file for a given uuid.
 
@@ -489,18 +364,7 @@ async def get_irods_root_collections():
     return {"folders": folders, "files": files}
 
 
-class CollectionItem(BaseModel):
-    path: Union[str, None] = None
-
-    class Config:
-        schema_extra = {
-            "example": {
-                "path": "/tempZone/home/rods/datasets",
-            }
-        }
-
-
-@ app.post("/collection", tags=["iRODS"])
+@ app.post("/collection", tags=["iRODS"], summary="Get folder information", response_description="All folders/files name and path under selected folder")
 async def get_irods_collections(item: CollectionItem):
     """
     Return all collections from the required folder.
@@ -519,13 +383,8 @@ async def get_irods_collections(item: CollectionItem):
                             detail="Data not found in the provided path")
 
 
-class Action(str, Enum):
-    preview = "preview"
-    download = "download"
-
-
-@ app.get("/data/{action}/{filepath:path}", tags=["iRODS"])
-async def get_irods_data_file(action: Action, filepath: str):
+@ app.get("/data/{action}/{filepath:path}", tags=["iRODS"], summary="Download irods file", response_description="A file with data")
+async def get_irods_data_file(action: ActionParam, filepath: str):
     """
     Used to preview most types of data files in iRODS (.xlsx and .csv not supported yet).
     OR
