+++ app/search.py
@@ -1,18 +1,12 @@
 import re
-
 from fastapi import HTTPException
 
+from app.data_schema import *
 from app.config import iRODSConfig
 
 from irods.column import Like, In
 from irods.models import Collection, DataObjectMeta
 
-
-BAD_REQUEST = 400
-UNAUTHORIZED = 401
-NOT_FOUND = 404
-METHOD_NOT_ALLOWED = 405
-INTERNAL_SERVER_ERROR = 500
 
 SEARCHFIELD = [
     "TITLE",
